handwritten-htr-bench/requirements.txt
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
pillow>=10.0
tensorflow>=2.14
tensorflow-datasets>=4.9
rich>=13.7
